requirements.txt
streamlit==1.28.1
requests>=2.32.0
folium==0.14.0
streamlit-folium==0.15.0
polyline==2.0.0
openai==0.28.1
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.25.2